fix(core): UBSan enum-invalid-value in vmaf_log and vmaf_option_set (ADR-1080)
Two tests trigger UBSan's enum-invalid-value check against production code
that correctly handles out-of-range inputs:

- test_log passes VMAF_LOG_LEVEL_NONE-1 (= 4294967295u as unsigned enum)
  to verify vmaf_log silently discards sub-zero severity levels. The
  comparison `level <= VMAF_LOG_LEVEL_NONE` at log.cpp:120 loads the
  invalid enum value through the enum type — UB under the C++ standard.

- test_opt passes (VmafOptionType)9999 to verify -EINVAL for unknown
  option types. The `switch(opt->type)` at opt.cpp:105 similarly loads
  an invalid enum value — UB.

Fix: cast to int before comparisons in vmaf_log and before the switch
in vmaf_option_set. The int cast is value-preserving for all valid and
invalid inputs; both enum types have int as their underlying type, so
no ABI or code-generation change results.

Fixes test_log and test_opt under the ASan+UBSan PR gate.

+++ core/src/opt.cpp
@@ -102,7 +102,14 @@ extern "C" [[nodiscard]] int vmaf_option_set(const VmafOption *opt, void *obj, c
 
     uint8_t *base = static_cast<uint8_t *>(obj) + opt->offset;
 
-    switch (opt->type) {
+    /* Cast to int before switching: test_dispatch_unknown_type passes 9999 as
+     * the opt->type field to verify -EINVAL is returned for unknown types.
+     * A switch on an enum whose value is not one of its named enumerators
+     * triggers UBSan's enum-invalid-value check.  Switching on int is
+     * semantically identical — the case labels are integer constants — and
+     * makes the dispatch UBSan-clean without changing behaviour.
+     * See ADR-1080 (UBSan enum-invalid-value in vmaf_log / vmaf_option_set). */
+    switch (static_cast<int>(opt->type)) {
     case VMAF_OPT_TYPE_BOOL: {
         bool *dst = reinterpret_cast<bool *>(base);
         *dst = opt->default_val.b;
